LPS-77418 Moving classes to the proper path
