+++ modules/apps/shopping/shopping-service/src/main/java/com/liferay/shopping/internal/upgrade/v1_0_1/util/ShoppingOrderTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.shopping.internal.upgrade.v1_0_2.util;
+package com.liferay.shopping.internal.upgrade.v1_0_1.util;
 
 import java.sql.Types;
