+++ modules/apps/shopping/shopping-service/src/main/java/com/liferay/shopping/upgrade/ShoppingServiceUpgrade.java
@@ -21,8 +21,8 @@ import com.liferay.shopping.upgrade.v1_0_0.UpgradeKernelPackage;
 import com.liferay.shopping.upgrade.v1_0_0.UpgradePortletId;
 import com.liferay.shopping.upgrade.v1_0_0.UpgradeShopping;
 import com.liferay.shopping.upgrade.v1_0_0.UpgradeShoppingPreferences;
-import com.liferay.shopping.upgrade.v1_0_1.UpgradeShoppingConfiguration;
-import com.liferay.shopping.upgrade.v1_0_2.UpgradeShoppingEmailAddress;
+import com.liferay.shopping.upgrade.v1_0_1.UpgradeShoppingEmailAddress;
+import com.liferay.shopping.upgrade.v1_0_2.UpgradeShoppingConfiguration;
 
 import org.osgi.service.cm.ConfigurationAdmin;
 import org.osgi.service.component.annotations.Component;
