+++ modules/apps/shopping/shopping-service/src/main/java/com/liferay/shopping/upgrade/v1_0_2/UpgradeShoppingConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.shopping.upgrade.v1_0_1;
+package com.liferay.shopping.upgrade.v1_0_2;
 
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
